# src/SamarPlanner.Web/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "SamarPlanner.Api",
    "version": "1.0"
  },
  "paths": {
...
    "/api/v1/tasks": {
      "post": {
        "tags": [
          "Task"
        ],
        "operationId": "CreateTask",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/CreateTaskRequest"
                  }
                ]
              }
            },
            "text/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/CreateTaskRequest"
                  }
                ]
              }
            },
            "application/*+json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/CreateTaskRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CreateTaskCommandResponseResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateTaskCommandResponseResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateTaskCommandResponseResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tasks/{taskId}": {
      "put": {
        "tags": [
          "Task"
        ],
        "operationId": "UpdateTask",
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/UpdateTaskRequest"
                  }
                ]
              }
            },
            "text/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/UpdateTaskRequest"
                  }
                ]
              }
            },
            "application/*+json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/UpdateTaskRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BooleanResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BooleanResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BooleanResult"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Task"
        ],
        "operationId": "DeleteTask",
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BooleanResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BooleanResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BooleanResult"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Task"
        ],
        "operationId": "GetTaskDetail",
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GetTaskDetailQueryResultResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetTaskDetailQueryResultResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetTaskDetailQueryResultResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tasks/{taskId}/soft-delete": {
      "put": {
        "tags": [
          "Task"
        ],
        "operationId": "SoftDeleteTask",
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BooleanResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BooleanResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BooleanResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tasks/{taskId}/restore": {
      "put": {
        "tags": [
          "Task"
        ],
        "operationId": "RestoreTask",
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BooleanResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BooleanResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BooleanResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tasks/with-occurrences": {
      "get": {
        "tags": [
          "Task"
        ],
        "operationId": "GetTasksWithOccurrences",
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GetTasksWithOccurrencesQueryResultResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetTasksWithOccurrencesQueryResultResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetTasksWithOccurrencesQueryResultResult"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "BooleanResult": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "isSuccess": {
            "type": "boolean",
            "readOnly": true
          },
          "badResultType": {
            "type": "string",
            "readOnly": true,
            "nullable": true,
            "enum": [
              "NotFound",
              "Validation",
              "General",
              "Unauthorized"
            ]
          },
          "response": {
            "type": "boolean",
            "readOnly": true
          },
          "errors": {
            "type": "object",
            "readOnly": true,
            "nullable": true,
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "CreateTaskCommandResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "CreateTaskCommandResponseResult": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "isSuccess": {
            "type": "boolean",
            "readOnly": true
          },
          "badResultType": {
            "type": "string",
            "readOnly": true,
            "nullable": true,
            "enum": [
              "NotFound",
              "Validation",
              "General",
              "Unauthorized"
            ]
          },
          "response": {
            "readOnly": true,
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/CreateTaskCommandResponse"
              }
            ]
          },
          "errors": {
            "type": "object",
            "readOnly": true,
            "nullable": true,
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
      "CreateTaskRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "title": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "defaultTime": {
            "type": "string",
            "format": "time",
            "nullable": true
          },
          "priority": {
            "type": "string",
            "nullable": true,
            "enum": [
              "High",
              "Medium",
              "Low"
            ]
          },
          "type": {
            "type": "string",
            "enum": [
              "Task",
              "Event"
            ]
          },
          "repeatPattern": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/RepeatPatternDto"
              }
            ]
          },
          "parentGoalId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          }
        }
      },
...
      "GetTaskDetailQueryResult": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "taskId": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "defaultTime": {
            "type": "string",
            "format": "time",
            "nullable": true
          },
          "priority": {
            "type": "string",
            "nullable": true,
            "enum": [
              "High",
              "Medium",
              "Low"
            ]
          },
          "type": {
            "type": "string",
            "enum": [
              "Task",
              "Event"
            ]
          },
          "repeatPattern": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/RepeatPatternDto"
              }
            ]
          },
          "parentGoalId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          }
        }
      },
      "GetTaskDetailQueryResultResult": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "isSuccess": {
            "type": "boolean",
            "readOnly": true
          },
          "badResultType": {
            "type": "string",
            "readOnly": true,
            "nullable": true,
            "enum": [
              "NotFound",
              "Validation",
              "General",
              "Unauthorized"
            ]
          },
          "response": {
            "readOnly": true,
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/GetTaskDetailQueryResult"
              }
            ]
          },
          "errors": {
            "type": "object",
            "readOnly": true,
            "nullable": true,
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
      "GetTasksWithOccurrencesQueryResult": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "tasks": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/GetTasksWithOccurrencesQueryResultTasks"
            }
          }
        }
      },
      "GetTasksWithOccurrencesQueryResultResult": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "isSuccess": {
            "type": "boolean",
            "readOnly": true
          },
          "badResultType": {
            "type": "string",
            "readOnly": true,
            "nullable": true,
            "enum": [
              "NotFound",
              "Validation",
              "General",
              "Unauthorized"
            ]
          },
          "response": {
            "readOnly": true,
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/GetTasksWithOccurrencesQueryResult"
              }
            ]
          },
          "errors": {
            "type": "object",
            "readOnly": true,
            "nullable": true,
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
      "GetTasksWithOccurrencesQueryResultTaskOccurrences": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "date": {
            "type": "string",
            "format": "date"
          },
          "time": {
            "type": "string",
            "format": "time",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "Pending",
              "Done",
              "NotDone",
              "AlmostDone"
            ]
          },
          "score": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "isSkipped": {
            "type": "boolean"
          }
        }
      },
      "GetTasksWithOccurrencesQueryResultTasks": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "taskId": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "priority": {
            "type": "string",
            "nullable": true,
            "enum": [
              "High",
              "Medium",
              "Low"
            ]
          },
          "type": {
            "type": "string",
            "enum": [
              "Task",
              "Event"
            ]
          },
          "parentGoalId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "occurrences": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/GetTasksWithOccurrencesQueryResultTaskOccurrences"
            }
          }
        }
      },
...
      "RepeatPatternDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "Daily",
              "WeeklyOnDays",
              "MonthlyOnDays"
            ]
          },
          "anchorDate": {
            "type": "string",
            "format": "date"
          },
          "interval": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "weekDays": {
            "type": "array",
            "nullable": true,
            "items": {
              "type": "integer",
              "format": "int32",
              "enum": [
                0,
                1,
                2,
                3,
                4,
                5,
                6
              ]
            }
          },
          "monthDays": {
            "type": "array",
            "nullable": true,
            "items": {
              "type": "integer",
              "format": "int32"
            }
          }
        }
      },
...
      "UpdateTaskRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "title": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "defaultTime": {
            "type": "string",
            "format": "time",
            "nullable": true
          },
          "priority": {
            "type": "string",
            "nullable": true,
            "enum": [
              "High",
              "Medium",
              "Low"
            ]
          },
          "type": {
            "type": "string",
            "enum": [
              "Task",
              "Event"
            ]
          },
          "repeatPattern": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/RepeatPatternDto"
              }
            ]
          },
          "parentGoalId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          }
        }
      }
    }
  },
...
}